# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ICC Procedural Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ApplicableRules =
  (typeof ApplicableRules)[keyof typeof ApplicableRules];

export const ApplicableRules = {
  ICC_2021: "ICC 2021",
  ICC_2017: "ICC 2017",
  ICC_2012: "ICC 2012",
} as const;

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  Active: "Active",
  Pending: "Pending",
  Closed: "Closed",
  Suspended: "Suspended",
} as const;

export type TribunalRole = (typeof TribunalRole)[keyof typeof TribunalRole];

export const TribunalRole = {
  President: "President",
  "Co-arbitrator": "Co-arbitrator",
  Sole_Arbitrator: "Sole Arbitrator",
  Secretary: "Secretary",
} as const;

export type RepresentativeRole =
  (typeof RepresentativeRole)[keyof typeof RepresentativeRole];

export const RepresentativeRole = {
  Lead_Counsel: "Lead Counsel",
  "Co-counsel": "Co-counsel",
  Junior: "Junior",
} as const;

export type RepresentativeParty =
  (typeof RepresentativeParty)[keyof typeof RepresentativeParty];

export const RepresentativeParty = {
  Claimant: "Claimant",
  Respondent: "Respondent",
} as const;

export interface Case {
  id: number;
  caseReference: string;
  caseName: string;
  claimants: string;
  respondents: string;
  seatOfArbitration: string;
  languageOfArbitration: string;
  applicableRules: ApplicableRules;
  dateOfRequest: string;
  currency: string;
  status: CaseStatus;
  createdAt: string;
  updatedAt: string;
}

export interface TribunalMember {
  id: number;
  caseId: number;
  name: string;
  role: TribunalRole;
  email: string;
  timeZone: string;
  createdAt: string;
}

export interface Representative {
  id: number;
  caseId: number;
  name: string;
  firm: string;
  role: RepresentativeRole;
  party: RepresentativeParty;
  email: string;
  timeZone: string;
  createdAt: string;
}

export type CaseDetail = Case & {
  tribunalMembers: TribunalMember[];
  representatives: Representative[];
};

export interface CreateCaseRequest {
  caseReference: string;
  caseName: string;
  claimants: string;
  respondents: string;
  seatOfArbitration: string;
  languageOfArbitration: string;
  applicableRules: ApplicableRules;
  dateOfRequest: string;
  currency: string;
  status?: CaseStatus;
}

export interface CreateTribunalMemberRequest {
  name: string;
  role: TribunalRole;
  email: string;
  timeZone: string;
}

export interface CreateRepresentativeRequest {
  name: string;
  firm: string;
  role: RepresentativeRole;
  party: RepresentativeParty;
  email: string;
  timeZone: string;
}

export type DeadlineStatus =
  (typeof DeadlineStatus)[keyof typeof DeadlineStatus];

export const DeadlineStatus = {
  Pending: "Pending",
  Completed: "Completed",
  Extended: "Extended",
} as const;

export type DeadlineResponsibleParty =
  (typeof DeadlineResponsibleParty)[keyof typeof DeadlineResponsibleParty];

export const DeadlineResponsibleParty = {
  Claimant: "Claimant",
  Respondent: "Respondent",
  Tribunal: "Tribunal",
  All: "All",
} as const;

export interface Deadline {
  id: number;
  caseId: number;
  description: string;
  responsibleParty: DeadlineResponsibleParty;
  dueDate: string;
  originalDueDate?: string | null;
  status: DeadlineStatus;
  proceduralOrderRef?: string | null;
  extensionOrderRef?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDeadlineRequest {
  description: string;
  responsibleParty: DeadlineResponsibleParty;
  dueDate: string;
  originalDueDate?: string | null;
  status: DeadlineStatus;
  proceduralOrderRef?: string | null;
  extensionOrderRef?: string | null;
  notes?: string | null;
}