# lib/api-zod/src/generated/types/deadline.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ICC Procedural Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { DeadlineResponsibleParty } from "./deadlineResponsibleParty";
import type { DeadlineStatus } from "./deadlineStatus";

export interface Deadline {
  id: number;
  caseId: number;
  description: string;
  responsibleParty: DeadlineResponsibleParty;
  dueDate: Date;
  originalDueDate?: Date | null;
  status: DeadlineStatus;
  proceduralOrderRef?: string | null;
  extensionOrderRef?: string | null;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}