# lib/api-zod/src/generated/types/deadlineResponsibleParty.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ICC Procedural Manager API
 * OpenAPI spec version: 0.1.0
 */

export type DeadlineResponsibleParty =
  (typeof DeadlineResponsibleParty)[keyof typeof DeadlineResponsibleParty];

export const DeadlineResponsibleParty = {
  Claimant: "Claimant",
  Respondent: "Respondent",
  Tribunal: "Tribunal",
  All: "All",
} as const;